Improving the webservice tests

+++ flaskblog/routesapi.py
@@ -1,7 +1,11 @@
-# file that implements the web service part of the project.
-# for more information about REST methods and their responsibilities,
-# visit: https://www.restapitutorial.com/lessons/httpmethods.html
-# or https://www.w3schools.in/restful-web-services/rest-methods/
+"""
+file that implements the web service part of the project.
+for more information about REST methods and their responsibilities,
+visit: https://www.restapitutorial.com/lessons/httpmethods.html
+or https://www.w3schools.in/restful-web-services/rest-methods/
+"""
+
+import sys
 from flask import request, jsonify, abort
 from flaskblog import app, db, bcrypt
 from flaskblog.models import Token, Post, User
@@ -9,9 +13,9 @@ import datetime
 
 
 # method used to create a token that can be used for some time defined by the delta
-@app.route('/api/token/public', methods=['GET'])
-def token():
-	data = request.json  # gets the JSON sent by the user
+@app.route('/api/token/public', methods=['POST'])
+def get_token():
+	data = request.form  # gets the JSON sent by the user
 
 	if 'email' not in data or 'password' not in data:
 		# in this case, we do not have enough information to perform a login
@@ -28,13 +32,14 @@ def token():
 			db.session.commit()
 			return jsonify({'token': token_string,
 							'message': 'Login successful!',
+							'user_id': user.id,
 							'expire': expired.strftime('%Y-%m-%d %H:%M:%S')})
 		except:
 			db.session.rollback()
 			return abort(400)  # HTTP code 400: bad request
 	else:
 		info = dict(message='Login Unsuccessful. Please check email and password.')
-		return jsonify(info),
+		return jsonify(info)
 
 
 # method used to inform the user of the webservice regarding its capabilities
@@ -67,18 +72,22 @@ def api_get_post(post_id):
 def api_create_post():
 	data = request.json  # gets the JSON sent by the user
 
+	token_string = request.headers['Authorization'].split(' ')[1]
+	token = Token.query.filter_by(token=token_string).first()
+
 	# the conditional should make sure that all the non-null attributes are present in the
 	# data sent by the call
-	if 'title' in data and 'content_type' in data and 'content' in data and 'user' in data:
+	if 'title' in data and 'content_type' in data and 'content' in data:
 		post = Post(title=data['title'],
 					content_type=data['content_type'],
 					content=data['content'],
-					user_id=int(data['user']))
+					user_id=token.user_id)
 		db.session.add(post)
 		try:
 			db.session.commit()
 			return jsonify(post), 201  # status 201 means "CREATED"
-		except:
+		except Exception as e:
+			print('The WebService API experienced an error: ', e, file=sys.stderr)
 			# to have more detailed exception messages, check the content of lecture 7
 			db.session.rollback()
 			abort(400)
@@ -92,13 +101,18 @@ def api_update_post(post_id):
 	post = Post.query.get_or_404(post_id)  # makes sure that the post_id exists
 	data = request.json
 
+	# verifying if the token used is of the user that is author of the post
+	token_string = request.headers['Authorization'].split(' ')[1]
+	cur_token = Token.query.filter_by(token=token_string).first()
+	if cur_token.user_id != post.user_id:
+		abort(401)
+
 	# the conditional should make sure that all the non-null attributes are present in the
 	# data sent by the call
 	if 'title' in data and 'content_type' in data and 'content' in data and 'user' in data:
 		post.title = data['title']
 		post.content_type = data['content_type']
 		post.content = data['content']
-		post.user_id = data['user']
 		try:
 			db.session.commit()
 			return jsonify(post), 200
@@ -116,8 +130,14 @@ def api_replace_post(post_id):
 	post = Post.query.get_or_404(post_id)
 	data = request.json
 
+	# verifying if the token used is of the user that is author of the post
+	token_string = request.headers['Authorization'].split(' ')[1]
+	cur_token = Token.query.filter_by(token=token_string).first()
+	if cur_token.user_id != post.user_id:
+		abort(401)
+
 	# you should have at least one of the columns to be able to perform an update
-	if 'title' in data or 'content_type' in data or 'content' in data or 'user' in data:
+	if 'title' in data or 'content_type' in data or 'content' in data:
 
 		# the conditionals below check each of the possible attributes to be modified
 		if 'title' in data:
@@ -126,8 +146,7 @@ def api_replace_post(post_id):
 			post.content_type = data['content_type']
 		if 'content' in data:
 			post.content = data['content']
-		if 'user' in data:
-			post.user_id = data['user']
+
 		try:
 			db.session.commit()
 			return jsonify(post), 200
@@ -142,6 +161,13 @@ def api_replace_post(post_id):
 @app.route('/api/post/<int:post_id>', methods=['DELETE'])
 def api_delete_post(post_id):
 	post = Post.query.get_or_404(post_id)
+
+	# verifying if the token used is of the user that is author of the post
+	token_string = request.headers['Authorization'].split(' ')[1]
+	cur_token = Token.query.filter_by(token=token_string).first()
+	if cur_token.user_id != post.user_id:
+		abort(401)
+
 	db.session.delete(post)
 	try:
 		db.session.commit()
